hw1/task2/requirements.txt
Biopython
matplotlib
ete3
